refactor: move and rename `src/prettyPrinter.ts` to `src/ast/ast-printer.ts`
Fixes #1420
