+++ CONTRIBUTING.md
@@ -176,7 +176,7 @@ Some other codegen tests are as follows:
 
 ### Pretty-printer and AST comparators
 
-The entry point to the Tact AST pretty-printer is [src/prettyPrinter.ts](./src/prettyPrinter.ts). It is going to be used for the Tact source code formatter once the parser keeps comments and other relevant information.
+The entry point to the Tact AST pretty-printer is [src/ast/ast-printer.ts](./src/ast/ast-printer.ts). It is going to be used for the Tact source code formatter once the parser keeps comments and other relevant information.
 
 The AST comparator is defined in [src/ast/compare.ts](./src/ast/compare.ts). This is useful, for instance, for static analysis tools which can re-use the Tact TypeScript API.
