+++ src/test/rename.spec.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import { join } from "path";
 import { AstRenamer } from "../ast/rename";
-import { prettyPrint } from "../prettyPrinter";
+import { prettyPrint } from "../ast/ast-printer";
 import { trimTrailingCR, CONTRACTS_DIR } from "./util";
 import * as assert from "assert";
 import { getParser } from "../grammar";
